requirements.txt
django==3.2.5
pillow==8.3.1
psycopg2==2.9.1
django-environ==0.4.5
django-ckeditor==6.1.0